chore: reorganize config into separate module
Initially the config was a subcommand and so it was mixed in with the commands. This separates it out so that we can begin the process of pulling it out into it's own crate.
