+++ cmd/crates/soroban-test/tests/it/config.rs
@@ -3,7 +3,8 @@ use soroban_test::{AssertExt, TestEnv};
 use std::{fs, path::Path};
 
 use crate::util::{add_key, add_test_id, SecretKind, DEFAULT_SEED_PHRASE};
-use soroban_cli::commands::network::{self, LOCAL_NETWORK_PASSPHRASE};
+use soroban_cli::commands::network;
+use soroban_cli::config::network::passphrase::LOCAL as LOCAL_NETWORK_PASSPHRASE;
 
 fn ls(sandbox: &TestEnv) -> Vec<String> {
     sandbox
