+++ cmd/soroban-cli/src/commands/cache/actionlog/ls.rs
@@ -1,7 +1,6 @@
 use clap::command;
 
-use super::super::super::config::locator;
-use crate::commands::config::data;
+use crate::config::{data, locator};
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
