+++ cmd/soroban-cli/src/commands/cache/path.rs
@@ -1,5 +1,4 @@
-use super::super::config::locator;
-use crate::commands::config::data;
+use crate::config::{data, locator};
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
