+++ cmd/soroban-cli/src/commands/contract/invoke.rs
@@ -23,15 +23,13 @@ use soroban_env_host::{
 
 use soroban_spec::read::FromWasmError;
 
-use super::super::{
-    config::{self, locator},
-    events,
-};
+use super::super::events;
 use crate::commands::txn_result::{TxnEnvelopeResult, TxnResult};
 use crate::commands::NetworkRunnable;
 use crate::get_spec::{self, get_remote_contract_spec};
 use crate::{
-    commands::{config::data, global, network},
+    commands::global,
+    config::{self, data, locator, network},
     rpc, Pwd,
 };
 use soroban_spec_tools::{contract, Spec};
