+++ cmd/soroban-cli/src/commands/keys/add.rs
@@ -1,6 +1,6 @@
 use clap::command;
 
-use super::super::config::{locator, secret};
+use crate::config::{locator, secret};
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
