+++ cmd/soroban-cli/src/commands/keys/generate.rs
@@ -1,9 +1,7 @@
 use clap::{arg, command};
 
-use crate::commands::network;
-
 use super::super::config::{
-    locator,
+    locator, network,
     secret::{self, Secret},
 };
