+++ cmd/soroban-cli/src/commands/network/mod.rs
@@ -1,19 +1,8 @@
-use std::str::FromStr;
+use clap::Parser;
 
-use clap::{arg, Parser};
-use phf::phf_map;
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
-use stellar_strkey::ed25519::PublicKey;
-
-use crate::{
-    commands::HEADING_RPC,
-    rpc::{self, Client},
-};
+use crate::rpc::{self};
 
 use super::config::locator;
-
-pub const LOCAL_NETWORK_PASSPHRASE: &str = "Standalone Network ; February 2017";
 
 pub mod add;
 pub mod container;
@@ -113,167 +102,5 @@ impl Cmd {
             }
         };
         Ok(())
-    }
-}
-
-#[derive(Debug, clap::Args, Clone, Default)]
-#[group(skip)]
-pub struct Args {
-    /// RPC server endpoint
-    #[arg(
-        long = "rpc-url",
-        requires = "network_passphrase",
-        required_unless_present = "network",
-        env = "STELLAR_RPC_URL",
-        help_heading = HEADING_RPC,
-    )]
-    pub rpc_url: Option<String>,
-    /// Network passphrase to sign the transaction sent to the rpc server
-    #[arg(
-        long = "network-passphrase",
-        requires = "rpc_url",
-        required_unless_present = "network",
-        env = "STELLAR_NETWORK_PASSPHRASE",
-        help_heading = HEADING_RPC,
-    )]
-    pub network_passphrase: Option<String>,
-    /// Name of network to use from config
-    #[arg(
-        long,
-        required_unless_present = "rpc_url",
-        env = "STELLAR_NETWORK",
-        help_heading = HEADING_RPC,
-    )]
-    pub network: Option<String>,
-}
-
-impl Args {
-    pub fn get(&self, locator: &locator::Args) -> Result<Network, Error> {
-        if let Some(name) = self.network.as_deref() {
-            if let Ok(network) = locator.read_network(name) {
-                return Ok(network);
-            }
-        }
-        if let (Some(rpc_url), Some(network_passphrase)) =
-            (self.rpc_url.clone(), self.network_passphrase.clone())
-        {
-            Ok(Network {
-                rpc_url,
-                network_passphrase,
-            })
-        } else {
-            Err(Error::Network)
-        }
-    }
-}
-
-#[derive(Debug, clap::Args, Serialize, Deserialize, Clone)]
-#[group(skip)]
-pub struct Network {
-    /// RPC server endpoint
-    #[arg(
-        long = "rpc-url",
-        env = "STELLAR_RPC_URL",
-        help_heading = HEADING_RPC,
-    )]
-    pub rpc_url: String,
-    /// Network passphrase to sign the transaction sent to the rpc server
-    #[arg(
-            long,
-            env = "STELLAR_NETWORK_PASSPHRASE",
-            help_heading = HEADING_RPC,
-        )]
-    pub network_passphrase: String,
-}
-
-impl Network {
-    pub async fn helper_url(&self, addr: &str) -> Result<http::Uri, Error> {
-        use http::Uri;
-        tracing::debug!("address {addr:?}");
-        let rpc_uri = Uri::from_str(&self.rpc_url)
-            .map_err(|_| Error::InvalidUrl(self.rpc_url.to_string()))?;
-        if self.network_passphrase.as_str() == LOCAL_NETWORK_PASSPHRASE {
-            let auth = rpc_uri.authority().unwrap().clone();
-            let scheme = rpc_uri.scheme_str().unwrap();
-            Ok(Uri::builder()
-                .authority(auth)
-                .scheme(scheme)
-                .path_and_query(format!("/friendbot?addr={addr}"))
-                .build()?)
-        } else {
-            let client = Client::new(&self.rpc_url)?;
-            let network = client.get_network().await?;
-            tracing::debug!("network {network:?}");
-            let uri = client.friendbot_url().await?;
-            tracing::debug!("URI {uri:?}");
-            Uri::from_str(&format!("{uri}?addr={addr}")).map_err(|e| {
-                tracing::error!("{e}");
-                Error::InvalidUrl(uri.to_string())
-            })
-        }
-    }
-
-    #[allow(clippy::similar_names)]
-    pub async fn fund_address(&self, addr: &PublicKey) -> Result<(), Error> {
-        let uri = self.helper_url(&addr.to_string()).await?;
-        tracing::debug!("URL {uri:?}");
-        let response = match uri.scheme_str() {
-            Some("http") => hyper::Client::new().get(uri.clone()).await?,
-            Some("https") => {
-                let https = hyper_tls::HttpsConnector::new();
-                hyper::Client::builder()
-                    .build::<_, hyper::Body>(https)
-                    .get(uri.clone())
-                    .await?
-            }
-            _ => {
-                return Err(Error::InvalidUrl(uri.to_string()));
-            }
-        };
-        let body = hyper::body::to_bytes(response.into_body()).await?;
-        let res = serde_json::from_slice::<serde_json::Value>(&body)
-            .map_err(|e| Error::FailedToParseJSON(uri.to_string(), e))?;
-        tracing::debug!("{res:#?}");
-        if let Some(detail) = res.get("detail").and_then(Value::as_str) {
-            if detail.contains("createAccountAlreadyExist") {
-                eprintln!("Account already exists");
-            }
-        } else if res.get("successful").is_none() {
-            return Err(Error::InproperResponse(res.to_string()));
-        }
-        Ok(())
-    }
-
-    pub fn rpc_uri(&self) -> Result<http::Uri, Error> {
-        http::Uri::from_str(&self.rpc_url).map_err(|_| Error::InvalidUrl(self.rpc_url.to_string()))
-    }
-}
-
-pub static DEFAULTS: phf::Map<&'static str, (&'static str, &'static str)> = phf_map! {
-    "local" => (
-        "http://localhost:8000/rpc",
-        "Standalone Network ; February 2017",
-    ),
-    "futurenet" => (
-        "https://rpc-futurenet.stellar.org:443",
-        "Test SDF Future Network ; October 2022",
-    ),
-    "testnet" => (
-        "https://soroban-testnet.stellar.org",
-        "Test SDF Network ; September 2015",
-    ),
-    "mainnet" => (
-        "Bring Your Own: https://developers.stellar.org/docs/data/rpc/rpc-providers",
-        "Public Global Stellar Network ; September 2015",
-    ),
-};
-
-impl From<&(&str, &str)> for Network {
-    /// Convert the return value of `DEFAULTS.get()` into a Network
-    fn from(n: &(&str, &str)) -> Self {
-        Self {
-            rpc_url: n.0.to_string(),
-            network_passphrase: n.1.to_string(),
-        }
     }
 }
