+++ cmd/soroban-cli/src/config/mod.rs
@@ -13,17 +13,14 @@ use crate::{
 
 use self::{network::Network, secret::Secret};
 
-use super::{keys, network};
-
 pub mod alias;
 pub mod data;
 pub mod locator;
+pub mod network;
 pub mod secret;
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
-    #[error(transparent)]
-    Identity(#[from] keys::Error),
     #[error(transparent)]
     Network(#[from] network::Error),
     #[error(transparent)]
