+++ cmd/soroban-cli/src/config/network/passphrase.rs
@@ -1,0 +1,4 @@
+pub const LOCAL: &str = "Standalone Network ; February 2017";
+pub const TESTNET: &str = "Test SDF Network ; September 2015";
+pub const FUTURENET: &str = "Test SDF Future Network ; October 2022";
+pub const MAINNET: &str = "Public Global Stellar Network ; September 2015";
