+++ cmd/soroban-cli/src/config/secret.rs
